Change all our int32_t to long
Well in the MIB it says Integer32, however net-snmp uses 'long' as data
type for those integers defined in netsnmp_vardata in
include/net-snmp/types.h and guess what fails on x86_64:
netsnmp_check_vb_type_and_size() …

(Maybe this would have work on a 32 bit machine?)

Note: the set request itself on the wire uses only as much bytes as
needed to put the actual value in, so for example one byte for the value
100, two bytes for the value 1000, three bytes for the value 100000, and
four bytes for the value 1000000000.

SNMP is a strange protocol!

+++ nstMIBObjects.c
@@ -148,7 +148,7 @@ handle_nstAgentSubagentObject(netsnmp_mib_handler *handler,
          */
     case MODE_SET_RESERVE1:
         ret = netsnmp_check_vb_type_and_size(requests->requestvb,
-											 ASN_INTEGER, sizeof(int32_t));
+											 ASN_INTEGER, sizeof(long));
         if (ret != SNMP_ERR_NOERROR) {
             netsnmp_set_request_error(reqinfo, requests, ret);
         }
